refactor: move login screen to index.js chore: remove login.js
